fix: clean up audience schema by removing unused code and moving types
- Remove unused AudienceQueryFiltersSchema and related constants (VISIBILITY, SYNC_STATUS, SORT_FIELDS, SORT_ORDER)
- Remove unused AudienceModelValidators helper functions
- Move AudienceStats type from schema to types folder following project conventions
- Update imports across codebase to reference types from correct location
- Retain only AudienceStatsSchema which is actively used for API responses

This cleanup removes legacy database-oriented filtering code that doesn't align with the current read-only Mailchimp API architecture.
